backend/requirements.txt
flask~=1.1.2
flask-cors
praw
requests~=2.25.1
nltk
google-cloud-language
ratelimit